separate ensure_addon_data_json to create... and get...

+++ client/ayon_usd/utils.py
@@ -16,22 +16,29 @@ DOWNLOAD_DIR = os.path.join(USD_ADDON_ROOT_DIR, "downloads")
 ADDON_DATA_JSON_PATH = os.path.join(DOWNLOAD_DIR, "ayon_usd_addon_info.json")
 
 
-def ensure_addon_data_json() -> dict:
-    """Ensure addon metadata JSON exists and return its content."""
-    os.makedirs(DOWNLOAD_DIR, exist_ok=True)
-
+def get_addon_data_json() -> dict:
+    """Get addon data JSON content as dict."""
     if os.path.exists(ADDON_DATA_JSON_PATH):
         with open(ADDON_DATA_JSON_PATH, "r") as json_file:
             return json.load(json_file)
+    return {}
+
+
+def create_addon_data_json_file():
+    """Create addon data JSON file if it doesn't exist."""
+    if os.path.exists(ADDON_DATA_JSON_PATH):
+        return
+
+    os.makedirs(DOWNLOAD_DIR, exist_ok=True)
 
     init_data = {
         "ayon_usd_addon_first_init_utc": str(
             datetime.now().astimezone(timezone.utc)
         )
     }
+
     with open(ADDON_DATA_JSON_PATH, "w") as json_file:
         json.dump(init_data, json_file)
-    return init_data
 
 
 def get_download_dir(create_if_missing=True):
@@ -75,7 +82,7 @@ def is_usd_lib_download_needed(settings: dict) -> bool:
     if not os.path.exists(usd_lib_dir):
         return True
 
-    addon_data_json = ensure_addon_data_json()
+    addon_data_json = get_addon_data_json()
     try:
         usd_lib_lake_fs_time_stamp_local = addon_data_json[
             "usd_lib_lake_fs_time_cest"
